refactor(server): split runtime config worker service
